CP-1213: Refactor Pool.join to use VM.export(metadata=true;all=true).
Now the pool-join logics is:
* first, create all the Hosts, PIFs, PBDs, Networks, SRs, VDIs if they do not already exists on the pool we are trying to join (remark: every objects appart networks are uniquely identified by their uuid; networks are identified by their name)
* the logics for making the network configuration homogenous is not there: JonL and JonD are working on updating the current logics so I didn't try to be too clever there, as everything will change soon and the current logics is quite complicated;
* finally, all the VMs metadata are exported on the current pool, and the imported on the other one.

Obviously, this changeset may have a big impact on regression tests as this is completly new codepath for the last bullet, and import/export of metadata are known to be a bit fragile. So hopefully, this change will make that code more robust without generating too many regressions ...

NB: the new code keeps the same asserts than the old one (appart for allowing pool-join when some VMs are suspended) but some of them can be removed if needed. As example:
* it would be possible to pool-join if we have running VMs (with no current operation)
* it would be possible to pool-join if our pool size > 1
* it would be possible to pool-join if we have shared storage