CP-1215: Add a new call, VM.make_template, to make a template from a VM object (which can be a normal VM, a snapshot, a checkpoint).
This call clones a VM object and remove the memory image of the cloned VM if it exists, and detach the VBDs if they are attached.
This call is needed because the requirements say that VM.clone on a checkpoint are not allowed, but we still need a way to create a template from a checkpoint.