CP-754, CP-765: (suspend PV-drivers domains without destroying the domain) CP-769, CP-770: (resume PV-drivers domains) CP-760: xapi glue to make checkpointing works
The checkpoint algorithm works as follow:

* suspend the VMs
* flush all the backend devices in dom0
* snapshot the VM, including its memory image
* resume the initial VM

That is simple to do in theory, but that causes a lot of pain due to concurrency problems. All the actions related to VBDs (ie, flushing devices, snapshots) are protected by a recursive lock which also protect the VM on which the VBDs are attached. As this lock need also to be taken while we suspend/resume the VM in order xal will not try to clean-up the domain, we can hit a bad dead-lock here. I tryied to bypass this locking mechanism in case of checkpoints, that should be safe normally but I rely on the stress tests to verify that it is working correctly ...

Moreover, some care has to be taken when exporting checkpoints, as per default we won't the exported file to contain a normal template (so wihtout memory image, with power-state=halted and the VBDs detached).
So there is a new option to VM.export, which is "preserve_power_state=true" which preserve the power state of the exported VM(s).