CA-30362: fix support escalation: multiple clones on EQL takes forever.
Make the SM backends responsible for setting the SR "dirty" flag (triggering an SR.scan) rather than always doing it. This allows backends to do their own thing i.e. in the case of EQL this means 'never scan because it can take 20 minutes' while in the case of LVHD this means 'scan after certain operations to trigger the GC'. Eventually we can remove the whole "dirty" concept entirely.