CA-31546: fix deadlock in reboot/shutdown handling code
The code has now been untangled and is structured as follows:

-----------------------------------------------------------+
Reboot                      | Shutdown                     |
in-guest       | in-dom0    | in-guest      | in-dom0      |
-----------------------------------------------------------+
     actions_after_{reboot,shutdown} for Egenera           |
-----------------------------------------------------------+
  separate queueing for API vs internally-generated        |
-----------------------------------------------------------+
                      locking                              |
-----------------------------------------------------------+

The in-guest bit is treated separately from the in-dom0 bit because the former blocks for ages and is very parallel unlike the latter. Queueing and locking is added as close to the entrypoints as possible -- rather than in the guts of the thing -- to make it more obviously fine and not prone to deadlocks.