CA-32071: convert one unsupported, untested mechanism for PCI passthrough into another unsupported, untested mechanisn for PCI passthrough where the latter works well enough to demo SRIOV while the former doesn't.
This should have no effect *unless* someone has set the pre-existing VM.other_config:pci key to activate PCI passthrough.

We switch from using the backend/frontend trees in xenstore on domain start to explicitly hotplugging the device in via qemu *after* the domain is unpaused.

We initiate a (slow) hot unplug from the migration progress callback (called when xenguest prints percentage progress to stderr) and then synchronise with it in the suspend callback. The progress threshold is adjustable by setting the VM.other_config:pci-hotunplug-time key but defaults to 0.8

In the migration receiver we replug the device back in after the domain is unpaused. Note this is not ideal behaviour but suffices for an SRIOV demo.