[DMC] When importing VMs which may have been exported on pervious XenServer versions, ensure the memory_dynamic_{min,max} have sensible values.
For Geneva-style XVAs we set dynamic_min=dynamic_max=static_max
For Rio-style XVAs we set dynamic_min=dynamic_max=static_max if the VM came from a previous version, otherwise we preserve the values.

We say a VM came from a previous version if (cheesy hack alert) it has not got a VM.parent field since this was added in Midnight Ride for checkpoint. If we end up adding a field for any reason for this feature then we should key off that instead for sanity.