[DMC] Add a new per-VM RRD ('memory_target') which records the current xenstore /local/domain/domid/memory/target value.
We watch the memory/target key from xal and call a new callback_memory_target when it changes. We stick these values into a cache in the xapi_guest_agent.ml (chosen because it already provides a convenient way to remove entries from the table when domains disappear) and then poll this cache from the RRD monitoring code.