CA-31782: Improve error reporting when the requested amount of memory cannot be freed. Account for the memory which the squeezer keeps to itself -- for the lowmem_emergency_pool and misc other memory allocations in the Host memory overhead.
When xapi requests an amount of memory is freed the squeezer aims to make this free plus a few extra MiB for the lowmem emergency pool. This extra few MiB was leaking back through the error messages so it looked like the amount of memory requested and the amount of memory "free" was larger than expected. Fix this by subtracting the added overhead from the error messages.

Since the squeezer has to keep a few MiB free, we now account for this in the XenAPI by having the squeezer (which starts before xapi) write the total to xenstore and then xapi picks this up in dbsync. This is then added to Host.memory_overhead.