CP-1302: Create tool to adjust dom0 memory limit
This reads in the xml from the /etc/xensource/installed-repos/*/XS-REPOSITORY
files and adds together the values of the "memory-requirement-mb"
attributes from the <repository> nodes.  (Default is 0.)  It then sets
the memory-target of dom0 to:

   min(static_max, total_memory_requirement_mb + 100mb + static_min)

This tool should be called once following installation and once again
for each package installed while dom0 is running.