CA-32680: Add xapi support for communicating with a persistent SM daemon over a socket
Existing python SR backends will continue to work as normal but this patch adds the ability to instead communicate with a persistent SM daemon. Xapi will check in /var/xapi/sm for the presence of a socket (called e.g. 'NFS' for an NFS SR backend) and will use this in preference to the usual backend if it exists.

(Original patch by Jonathan Ludlam, modified by me.)