CA-32765: Improve event notifications
Keep track of the generation count of the most recent update which each client has been notified about, and then send him all the changes since that generation count upon the next call to Event.next().

Note that, whilst there's no change to the type of Event.next(), this patch changes its semantics: the values of objects which have been deleted no longer get sent to the client. If someone using the API starts complaining that they no longer see the values of deleted objects in the events, this is the patch they have a problem with. (The reason this patch doesn't change the type of Event.next() is that the `snapshot' field of an event has always been possibly-empty.

(Original patch by Jonathan Ludlam, modified by Jonathan Davies.)