Renaming the field name from 'secret' to 'value'.
Apparently the tools used to generate bindings in api-bindings.hg require that
the object name is different from all field names of the object.  'secret'
violated this by having a field named 'secret'.

It was deemed easier to fix api.hg rather than address the limitation in the
tools used to generate the bindings.