CA-33452: fix VM.start on slaves during incomplete rolling upgrade, for ease of testing.
Although we only guarantee to make the minimum functionality work during rolling upgrade for the upgrade to complete (i.e. shutdown, suspend, migrate), in practice it's a pain to debug a rolling upgrade migration issue if you cannot also start the VM to try a second attempt.

Remove the new Host.memory_total field and use the Host_metrics.memory_total instead. Make sure the metrics exist across all 3 cases:
1. fresh install: created via dbsync_slave -> Xapi_host.create
2. pool.join: Host.create -> Xapi_host.create
3. upgrade: created via dbsync.ml

(3) can be removed after this upgrade; it's only needed because of the PersistNone option on the Host_metrics table...

Signed-off-by: David Scott <dave.scott@eu.citrix.com>