CA-32810: prevent the memory ballooning daemon capping a domain's memory usage before it has written feature-balloon.
Initially the ballooning daemon treated all domains which had not yet started running as using their "initial-reservation". This relied on the assumption that the initial memory target would be reached by running the domain builder. This assumption is false because some memory accounted to the domain in totpages (eg VGA frame buffer) is allocated dynamically after unpause.

Instead we now consider all domains which have not yet written "feature-balloon" to be using the max of "initial-reservation" and totpages. This relies on the new, much weaker, assumption that, by the time "feature-balloon" is written, all memory allocations have ceased. Note that we never change a balloon target until "feature-balloon" is written so the domain will still have target=startmem at this point.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>