CA-35372: when setting maxmem to cap the allocations of a domain, set the target too.
This required rewriting the logic which determines whether domains are active or inactive.
Now we say that a domain is 'making progress' if it is moving towards its target. Domains
which are 'making progress' are always active. We say that a domain is inactive if (i) it
has not reached its target; and (ii) it hasn't been 'making progress' for more than the
threshold (currently 5s).

Signed-off-by: David Scott <dave.scott@eu.citrix.com>