[ocamldoc] Grouping modules in the module list
Introduced a custom tag "@group" to ocamldoc. Modules having the same @group tag in the first ocamldoc comment of the module are grouped together under the same heading in the module list.

I started adding @group tags to the modules in the xapi executable. This is not finished yet, and I am hoping someone will help me to finish this task and iron out inconsistencies... ;)

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>