[ocamldoc] Information on the completeness of the documentation
This patch introduces percentages at modules and components on the ocamldoc pages, indicating how "complete" the documentation is. Completeness is defined as the percentage of the elements (modules, functions, types, ...) that have a proper ocamldoc description. I'll only remove these statistics when there are all "100%" ;)

The xapi toolstack docs are now 12% complete!

Note that even though many descriptions are still missing, the structure of the modules and their dependencies can already be explored.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>