CA-36074: fix the udev SR
Unfortunately the changeset containing the updates to use the new fork/exec protocol accidentally included a change which is part of the experimental 'vhdd' project. The change would be harmless except for the fact that the udev SR is 'created' rather than 'introduced' by the firstboot script and the backend doesn't currently support probe.

Comment out the offending code for now.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>