CA-36075: add back in a sane default $PATH for subprocesses that need it.
In particular the 'EXTSR' storage backend expects to find 'pvcreate' on the path. Rather than just fixing this one instance and then seeing what else fails, this patch should prevent this class of error recurring.

Note that the only place we actually pass a custom environment to a subprocess is when installing the debian etch template.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>