[rpc-light] when (un)marshaling variant, if it has no arguments then consider it as a string.
This bit is also necessary to discuss with the SM backend.
Basically, if you have 'type t = Foo | Bar of int with rpc' you will consider than the value Foo is actually the same thing as the string "Foo" (if you don't want to have a capital letter, use polymorphic variants as 'type t = [ `foo | `bar of int ]' which will give that the value `foo will be considered as the string "foo").

Signed-off-by: Thomas Gazagnaire <thomas.gazagnaire@citrix.com>