CA-36294: Fix performance issues when credentials are given using the pool secret.
Previously, to check whether the received pool-secret was valid or not, RBAC was trying to do a proper login using Session.slave_login and check whether that call succeed or not; This makes DB calls on slave to pass from around 2ms (before RBAC changes) to around 10ms (now). When doing a VM.start, slaves are doing hundreds of DB access, so VM.start were becoming very slow (up to 5 times slower!).

This patch simply compare the pool secret given in the request with the one stored inside xapi memory which improve greatly the performance of slaves' calls.

Issue found with Jonathan Davies' great help.

Signed-off-by: Thomas Gazagnaire <thomas.gazagnaire@citrix.com>