CA-36316: fix a bug where the domain's maxmem gets out of sync in the case when no ballooning is required e.g. if dynamic_min = dynamic_max. This causes subsequent suspend/resume or migrate operations to demand more memory than they strictly need.
The problem is caused by:
1. xapi creates the domain and sets maxmem = static_max
2. squeezed modifies maxmem only if it is doing some squeezing; if dynamic_min = dynamic_max then no squeezing is necessary
3. xapi uses max(memory_actual, maxmem) as the amount "needed" by suspend/resume/migrate

Since squeezed is managing maxmem, this patch makes it always reset maxmems to low values in its polling loop. Note that there is a cache so we don't issue more hypercalls than we actually need.

Also fix a logging glitch where if dynamic_min = dynamic_max, large negative numbers were written to the log (but not used for anything)

Signed-off-by: David Scott <dave.scott@eu.citrix.com>