CA-38567: Catch I/O errors during Debian postinstall.
This somewhat redefines scripted postinstall code to exit with kernel
error numbers when failing VDI provisioning.

Partially works for ENOSPC (swap/dd only). Works for tar (rootfs) or
dd (swap) failing with I/O errors. Was surprised to learn gnu tar
apparently never fsyncs, it took an -osync mount to let it learn that
not all is great in storage land.

Unexpected return codes now raise an internal_error, not a wild bet
on provision_failed_out_of_space, as it used to.

Signed-off-by: Daniel Stodden <daniel.stodden@citrix.com>