Cleanup creation of Linux templates.
The way bug workarounds is applied, particularly to the RHEL and
related templates, is pretty confusing and leads to errors (for
example RHEL 5.4 does not suffer fro the machine address size bug but
still has the workaround applied). Instead of creating a new function
for each point release which requires different workarounds to its
predecessor instead define one function per major release which takes
flags to indicate which workaround should be applied.

I have confirmed that the created templates are identical before and
after this change, modulo the key ordering on the other-config
map. Fixes to which workarounds are actually enabled will follow in a
separate patch.

Signed-off-by: Ian Campbell <ian.campbell@citrix.com>