CP-1591: MTU param in xe network-create and vif-create
Since the Network object now has a read/write MTU param, we would like to set it when creating the network. This patch adds this option to the CLI. The VIF.MTU field is supposed to reflect the MTU of the Network it is on, and should not be set directly. Therefore, the MTU parameter of xe vif-create has been removed, and the MTU of the VIF object will be set to the MTU of its Network.

pif/vif plug/unplug

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>