CA-38944: prevent the artificial-reboot-delay being accidentally applied to an API-driven reboot.
When the request queueing was fixed, a deliberate race was created between the background event thread and the API VM.reboot (etc). One of these threads wins the race and then the second thread becomes a no-op. Unfortunately the background event thread will apply the artificial-reboot-delay unless it is somehow prevented. We prevent it by storing the next reboot delay in xenstore (rather than the database) and setting it to 0 in the API call path.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>