CA-38109: HostInstallation/TCQuicktest: quicktest failed: SR_OPERATION_NOT_SUPPORTED: OpaqueRef:2c96e6a1-afdc-19a6-b6cb-37af7e90e504 CP-1603: CR-50: R4. Include REQ259 - Writable ISO SRs
Update quicktest to match new writable ISO SR support.  The XenServer
Tools SR is special-cased to prevent writing even if the ISO SR backend
itself allows it, and quicktest needed to be updated to match this.

Also, the VM import code has a special case to protect importing into
ISO SRs, so quicktest needs to know about this, too.

Signed-off-by: Ewan Mellor <ewan.mellor@eu.citrix.com>