Add CPUID features whitelisting foor Pool.join.
This allows the user to exclude CPUID features from the Pool.join checks by setting a mask in Pool.other_config:cpuid_feature_mask. When comparing the CPUID features of the pool and the joining host for equality, this mask is applied before the comparison. The format is the same as the format of the feature field in Host.cpu_info. The default is "ffffff7f-ffffffff-ffffffff-ffffffff", which defines the EST feature, bit 7 of the base ecx flags, as "don't care".

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>