CA-37151: If a SIGTERM/SIGINT is received during metadata backup we should tidy up properly
This means run cleanup() if the signal is received and in cleanup block the signal so that
a second SIGTERM/SIGINT doesn't cause a problem.

Also, in cleanup(), if we have an uninitialised vdi (e.g. because we've just created it but
have not yet mkfs'd it) then we ought to delete the vdi.

Signed-off-by: Alex Zeffertt <alex.zeffertt@eu.citrix.com>