[PATCH] xenserver: Hoist identical bridge and vswitch functions into common code.
From 96c7918c4043326bfa0e05b3fc9cbb9547279d40 Mon Sep 17 00:00:00 2001
Date: Tue, 23 Feb 2010 09:47:31 -0800
The previous commit made pif_bridge_name() in the bridge and vswitch
versions of interface-reconfigure functionally identical, so this commit
hoists them into a single common implementation in InterfaceReconfigure.py.

pif_is_bridged() also comes along for the ride because it is also generic
and because it is logically related.  Only the bridge code uses it at the
moment.

Suggested-by: Ian Campbell <ian.campbell@citrix.com>
Signed-off-by: Ben Pfaff <blp@nicira.com>
Signed-off-by: Ian Campbell <ian.campbell@citrix.com>
---
 .../opt_xensource_libexec_InterfaceReconfigure.py  |   28 ++++++++++++++++++++
 ...xensource_libexec_InterfaceReconfigureBridge.py |   28 --------------------
 ...ensource_libexec_InterfaceReconfigureVswitch.py |   16 -----------
 3 files changed, 28 insertions(+), 44 deletions(-)