[PATCH] xenserver: Set internal network-uuids in Bridge table on XS5.5
From 79eb8b002c588e6f7495a0cd73d8a0597a458442 Mon Sep 17 00:00:00 2001
Date: Wed, 21 Apr 2010 03:42:52 -0700
On XenServer 5.5, interface-reconfigure is not called when creating
internal bridges, so we jump through extra hoops to determine the
network UUIDs.  The code that handled this was not properly retrieving
the UUIDs from XAPI, so the field would never be set.  This commit
corrects that.

Bug #2666

Signed-off-by: Ian Campbell <ian.campbell@citrix.com>
---
 scripts/vif |    8 ++------
 1 files changed, 2 insertions(+), 6 deletions(-)