[PATCH] interface-reconfigure: callout to datapath backend class method on rewrite
From 823c5699d4de7bf726f988e1ca6197fb2400f388 Mon Sep 17 00:00:00 2001
Date: Fri, 4 Jun 2010 16:39:10 +0100
Use this mechanism to allow the vswitch backend to update the vswitch
configuration's mapping from datapath to XenAPI datamodel Network
UUIDs. The vswitch needs a mechanism to update these when they change
(i.e. on pool join and eject).

Refactor the DatapathFactory method to return the class which the
caller can instantiate or not as the require.

Signed-off-by: Ian Campbell <ian.campbell@citrix.com>
Signed-off-by: Dominic Curran <dominic.curran@citrix.com>
---
 .../opt_xensource_libexec_InterfaceReconfigure.py  |   12 +++++++++---
 ...ensource_libexec_InterfaceReconfigureVswitch.py |   11 +++++++++++
 .../opt_xensource_libexec_interface-reconfigure    |    9 ++++++---
 3 files changed, 26 insertions(+), 6 deletions(-)