Coalesce RHEL/CentOS/OEL 5.x templates into one template per distro per arch
Create 6 new templates (RHEL,CentOS,OEL)*(i386,x86_64) that can be used to
install 5.x point release.

For now we leave the old templates in place to prevent a failure in the
regression testing.  However, when XenRT has been updated to use the new
templates we shall remove the old ones.

Signed-off-by: Alex Zeffertt <alex.zeffertt@eu.citrix.com>