MAR-112: RRD transfer lock can prevent master transition
If the master loses connectivity during a RRD transfer, the RRD mutex
will never be released.  When performing a emergency master transition,
the first thing that occurs is to save the local RRDs to disk, which
requires the RRD mutex.  This will never succeed, thus XAPI gets stuck
trying to shut down.

Instead of locking the mutex, we try to acquire it for a few seconds.
If the lock can not be taken, the loss of some RRD data is preferable
to not being able to transition at all.

Signed-off-by: Ted M Lin <tlin@marathontechnologies.com>