CA-41832: clip the target of an 'inactive' domain to be within dynamic_min/dynamic_max
This is potentially dangerous: the reason we attempt to cap the memory usage of an unresponsive domain is to prevent it suddenly waking up and allocating at a bad moment, resulting in either a transient OOM or loss of low memory. With this change the cap will be 'loose', where target can be higher than memory_actual.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>