CP-1591: Synchronise VIF.MTU and PIF.MTU on plug
The VIF.MTU and PIF.MTU fields are now read-only, and will reflect the current state of the network. When you change the Network.MTU, VIFs and PIFs will have to be replugged to have their MTU changed by the networking subsystem. So when Network.MTU and PIF.MTU are not equal, this is a reminder to replug the PIF (same for VIFs).

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>