CA-27648: xe does not show the tags fields --- FIXED.
Added the ability to read and write tags on pool, host, network, VM, SR, and
VDI through the CLI. For each corresponding function, the patch adds four
lines of the general form:

	make_field ~name:"tags"
		~get:(fun () -> String.concat ", " (x ()).API.xxx_tags)
		~get_set:(fun () -> (x ()).API.xxx_tags)
		~add_to_set:(fun tag -> Client.Xxx.add_tags rpc session_id xx tag)
		~remove_from_set:(fun tag -> Client.Xxx.remove_tags rpc session_id xx tag) ();

The patch also includes code re-indentation for the functions involved.

It appears as if much of this file could be auto-generated.

Signed-off-by: Rok Strnisa <rok.strnisa@citrix.com>